ml-service/requirements.txt
setuptools
flask==3.0.3
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-flask==0.48b0
opentelemetry-exporter-otlp==1.27.0
